@@ -2,7 +2,7 @@
   "name": "x4codecomplete",
   "displayName": "X4CodeComplete",
   "description": "",
-  "version": "1.0.2",
+  "version": "1.1.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/Cgettys/X4CodeComplete"
@@ -45,14 +45,19 @@
     "test": "npm run compile && node ./node_modules/vscode/bin/test"
   },
   "devDependencies": {
-    "typescript": "^3.3.1",
-    "vscode": "^1.1.28",
-    "tslint": "^5.12.1",
-    "@types/node": "^10.12.21",
-    "@types/mocha": "^2.2.42"
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^22.10.2",
+    "@types/xml2js": "^0.4.14",
+    "minimatch": "^10.0.1",
+    "minimist": "^1.2.8",
+    "mkdirp": "^3.0.1",
+    "mocha": "^11.0.1",
+    "tslint": "^5.20.1",
+    "typescript": "^5.7.2",
+    "vscode": "^1.1.34"
   },
   "dependencies": {
-    "xml2js": "^0.4.19",
-    "xml2js-xpath": "^0.10.0"
+    "xml2js": "^0.6.2",
+    "xml2js-xpath": "^0.13.0"
   }
 }
